requirements/devel.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file devel.txt devel.in
#
amqp==2.2.2               # via kombu
billiard==3.5.0.3         # via celery
celery==4.1.0
click==6.7                # via pip-tools
django-celery-beat==1.0.1
django-celery-results==1.0.1
django==1.11.5
djangorestframework==3.6.4
first==2.0.1              # via pip-tools
flake8==3.4.1
kombu==4.1.0              # via celery
mccabe==0.6.1             # via flake8
pip-tools==1.10.1
pycodestyle==2.3.1        # via flake8
pyflakes==1.5.0           # via flake8
pytz==2017.2              # via celery, django
six==1.11.0               # via pip-tools
vine==1.1.4               # via amqp
